refactor: remove helmut.misc.collection
